Add new post type and widget to build post project
Add new option for editor widget for better control

Add also new widget post in grid to show post like a card with image and
title, and excerpt if need, all data is take from the post is it call,
some option to skip the post if the widget is call from a post type.

The post type widget if the layput for all project type, take title for
post title, use a editor, and other widget for complete layout, for
image use a hack to to a 16:9 ratio.